backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# LangChain ecosystem
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
langchain-huggingface==0.0.3

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.6.0

# Document processing
PyMuPDF==1.23.21

# Web search
duckduckgo-search==4.4.3

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1
uuid6==2024.1.12
